Refactor: Remove unused components, hooks, and services
- Deleted Toaster component and related toast hooks to streamline UI notifications.
- Removed mock data for scores and test results to clean up the data layer.
- Eliminated unused socket connection logic and related hooks.
- Cleaned up service files by removing unused alert and profile services.
- Removed deprecated auto-save and debounce hooks.
- Deleted unnecessary context providers and Redux slices.
- Updated axios configuration to use environment variables for API base URL.
- Added folder animation utilities for improved UI transitions.
- Introduced new hooks for managing chat threads and auto-saving functionality.
