+++ .env.sample
@@ -1,0 +1,3 @@
+VITE_API_BASE_URL=
+VITE_SOCKET_URL=
+VITE_OPEN_AI_KEY=
