+++ index.html
@@ -8,17 +8,6 @@
     <link href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@700&display=swap" rel="stylesheet" />
   </head>
   <body>
-    <!-- index.html -->
-    <!-- <div id="splash-screen" style="
-      background-color: #13130e;
-      height: 100vh;
-      width: 100vw;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      font-family: sans-serif;
-    ">
-    </div> -->
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
