+++ src/app/pages/auth/root.tsx
@@ -1,6 +1,4 @@
-import { Outlet, useNavigate, 
-  // useNavigate 
-} from "react-router-dom";
+import { Outlet, useNavigate } from "react-router-dom";
 import AuthLayout from "@/components/layouts/auth";
 import { useGetCurrentUser } from "@/service/auth/get-current.user";
 import { useEffect } from "react";
@@ -17,15 +15,12 @@ const AuthRoot = () => {
 
   useEffect(() => {
     if(!user){
-      console.error("no state user");
       navigate("/auth/login");
       return;
     }
     if (data) {
-      console.log("has data, navigating to dashboard", data);
       navigate("/");
     } else if (error) {
-      console.error("no data or error, stay on login");
       navigate("/auth/login");
     }
   }, [data, error, navigate, user]);
