+++ src/app/pages/folder/folder-animation-utils.ts
@@ -1,0 +1,17 @@
+export const folderContainerVariants = {
+        hidden: { opacity: 0 },
+        visible: {
+        opacity: 1,
+        transition: { staggerChildren: 0.1 },
+    },
+}
+  
+export const folderItemVariants = {
+    hidden: { y: 20, opacity: 0 },
+    visible: {
+        y: 0,
+        opacity: 1,
+        transition: { type: "spring", stiffness: 100 },
+    },
+}
+
