+++ src/app/splash.tsx
@@ -1,5 +1,3 @@
-
-
 export default function SplashScreen() {
     return (
         <div className="w-screen h-screen flex items-center justify-center bg-background text-on-background">
