+++ src/components/search/search-modal.tsx
@@ -6,7 +6,7 @@ import { RootState } from "@/redux/store"
 import { useSelector } from "react-redux"
 import { Note } from "@/types/note"
 import { Folder } from "@/types/folder"
-import { useDebounce } from "@/hooks/useDebounce"
+import { useDebounce } from "@/hooks/use-debounce"
 import { getUserNotes } from "@/service/notes/get-user-notes"
 import SearchModalSkeleton from "./search-modal-skeleton"
 import SearchNoteItem from "./search-note-item"
