--- src/data/results.ts
@@ -1,76 +1,0 @@
-
-// import { TestResult } from '@/types/test-result';
-// import { v4 as uuidv4 } from 'uuid'; // If you're using uuid for unique ids
-
-// export const testResults: TestResult[] = [
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567890",
-//     testName: "Complete Blood Count",
-//     size: 5.2,
-//     createdAt: new Date("2024-07-01T10:00:00Z"),
-//   },
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567891",
-//     testName: "Lipid Profile",
-//     size: 3.8,
-//     createdAt: new Date("2024-07-02T09:30:00Z"),
-//   },
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567892",
-//     testName: "Thyroid Panel",
-//     size: 4.1,
-//     createdAt: new Date("2024-07-03T11:15:00Z"),
-//   },
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567893",
-//     testName: "Urine Test",
-//     size: 1.7,
-//     createdAt: new Date("2024-07-04T08:45:00Z"),
-//   },
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567894",
-//     testName: "Liver Function Test",
-//     size: 6.0,
-//     createdAt: new Date("2024-07-05T10:20:00Z"),
-//   },
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567895",
-//     testName: "Kidney Function Test",
-//     size: 5.4,
-//     createdAt: new Date("2024-07-06T13:10:00Z"),
-//   },
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567896",
-//     testName: "Glucose Test",
-//     size: 2.5,
-//     createdAt: new Date("2024-07-07T14:05:00Z"),
-//   },
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567897",
-//     testName: "Vitamin D",
-//     size: 3.2,
-//     createdAt: new Date("2024-07-08T12:30:00Z"),
-//   },
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567898",
-//     testName: "Iron Panel",
-//     size: 4.7,
-//     createdAt: new Date("2024-07-09T15:40:00Z"),
-//   },
-//   {
-//     id: uuidv4(),
-//     userPid: "U1234567899",
-//     testName: "COVID-19 PCR",
-//     size: 2.0,
-//     createdAt: new Date("2024-07-10T16:50:00Z"),
-//   }
-// ];
