+++ src/lib/chat-thread-manager.ts
